requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
PyQt5>=5.15.0
SpeechRecognition>=3.10.0
gtts>=2.4.0
pygame>=2.5.0
cvzone>=1.6.0
pyaudio>=0.2.13
google-generativeai>=0.3.0
